[PC-13325][API] Move native API Pydantic BaseModel to general API
Since we need the BaseModel for all the APIs, let's move it to
the base location.
