+++ api/src/pcapi/routes/native/v1/serialization/authentication.py
@@ -1,8 +1,7 @@
 from typing import Optional
 
+from pcapi.routes.serialization import BaseModel
 from pcapi.serialization.utils import to_camel
-
-from . import BaseModel
 
 
 class SigninRequest(BaseModel):
