+++ api/src/pcapi/routes/native/v1/serialization/favorites.py
@@ -8,9 +8,8 @@ from pcapi.core.categories.subcategories import SubcategoryIdEnum
 from pcapi.core.offers.api import get_expense_domains
 from pcapi.core.users.models import ExpenseDomain
 from pcapi.routes.native.utils import convert_to_cent
+from pcapi.routes.serialization import BaseModel
 from pcapi.utils.date import format_into_utc_date
-
-from . import BaseModel
 
 
 class Coordinates(BaseModel):
