+++ api/src/pcapi/routes/native/v1/serialization/offers.py
@@ -23,10 +23,9 @@ from pcapi.domain.show_types import SHOW_SUB_TYPES_DICT
 from pcapi.domain.show_types import SHOW_TYPES_DICT
 from pcapi.routes.native.utils import convert_to_cent
 from pcapi.routes.native.v1.serialization.common_models import Coordinates
+from pcapi.routes.serialization import BaseModel
 from pcapi.serialization.utils import to_camel
 from pcapi.utils.date import format_into_utc_date
-
-from . import BaseModel
 
 
 logger = logging.getLogger(__name__)
