+++ api/src/pcapi/routes/native/v1/serialization/subscription.py
@@ -5,11 +5,10 @@ from pydantic import validator
 from pcapi.core.subscription import models as subscription_models
 from pcapi.core.subscription import profile_options
 from pcapi.core.users import models as users_models
+from pcapi.routes.serialization import BaseModel
 from pcapi.serialization.utils import has_latin_or_numeric_chars
 from pcapi.serialization.utils import is_latin
 from pcapi.serialization.utils import to_camel
-
-from . import BaseModel
 
 
 class NextSubscriptionStepResponse(BaseModel):
